+++ models/word_analyzer.py
